requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
db-dtypes==1.0.5
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-cloud-bigquery==3.4.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.14
pandas==1.5.3
gspread==5.7.2